perf(address): compare recent-calls count as usize, not via usize->u32 cast (review #81)
`recent.len() as u32 >= CONTRACT_CALL_LIMIT` cast the usize len down to u32.
Harmless today (the query's LIMIT bounds len well under u32::MAX), but the cast
is the wrong direction. Compare in usize by casting the small constant up:
`recent.len() >= CONTRACT_CALL_LIMIT as usize`.

+++ src/network/address.rs
@@ -4022,7 +4022,7 @@ pub(super) async fn fetch_address_contract_calls(
         .await
     {
         Ok(recent) => {
-            let filled = recent.len() as u32 >= CONTRACT_CALL_LIMIT;
+            let filled = recent.len() >= CONTRACT_CALL_LIMIT as usize;
             info!(
                 addr = %format!("{:#x}", address),
                 recent_calls = recent.len(),
